% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcTSP3.R
\name{extractArtemis2009FormatFromTSP3ForMetaModelling}
\alias{extractArtemis2009FormatFromTSP3ForMetaModelling}
\title{Extract plot list for Artemis simulation}
\usage{
extractArtemis2009FormatFromTSP3ForMetaModelling(QcTSP3Data, plots)
}
\arguments{
\item{QcTSP3Data}{the database that is retrieved through the restoreQcTSP3Data function}

\item{plots}{a vector of integers standing for the plot id to be considered}
}
\value{
a data.frame object formatted for Capsis Web API
}
\description{
Extract plot list for Artemis simulation
}
