% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcTSP3.R
\name{extractNatura2014FormatFromTSP3ForMetaModelling}
\alias{extractNatura2014FormatFromTSP3ForMetaModelling}
\title{Extract plot list for Natura simulation}
\usage{
extractNatura2014FormatFromTSP3ForMetaModelling(QcTSP3Data, stratumPlots)
}
\arguments{
\item{QcTSP3Data}{the database that is retrieved through the restoreQcTSP3Data function}

\item{stratumPlots}{a dataframe with a column "stratum" as a grouping variable as a character field and a column "plots" for plots number as an integer}
}
\value{
a list of two dataframes where the first is a tree list and the second is a list of study trees for model Natura on Capsis Web API
}
\description{
Extract plot list for Natura simulation
}
