% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QcTSP3.R
\name{extractArtemisFormatFromTSP3}
\alias{extractArtemisFormatFromTSP3}
\title{Extract plot list for Artemis simulation}
\usage{
extractArtemisFormatFromTSP3(QcTSP3Data, plots, format = "WebAPI")
}
\arguments{
\item{QcTSP3Data}{the database that is retrieved through the restoreQcTSP3Data function}

\item{plots}{a vector of integers standing for the plot id to be considered}

\item{format}{a character string identifying the intended use, either "WebAPI" or "RPackage"}
}
\value{
a data.frame object formatted for Capsis Web API
}
\description{
Extract plot list for Artemis simulation
}
\examples{
\dontrun{
 extractArtemisFormatFromTSP3(QcTSP4Data=Data, plots= c(8500101, 9400909104,  9904805304),
  format="WebAPI")
}
}
